refactor: Move elf-specific symbol sizing logic
